Add data pipeline for MNIST dataset
This commit adds a rudimentary data pipeline for the MNIST dataset. If
the dataset is not installed on the local machine, it will be
downloaded, extracted and stored as a tuple of numpy arrays. If the
dataset should already be present, the previously stored .npy file is
loaded from disk.

As of now, no preprocessing is performed. This may change in the future.