Conduct Experiments on the Caltech-101 dataset
This commit adds experiments with capsule networks on the Caltech-101
dataset. In particular, this commit introduces two additions to the
codebase:

1. a data pipeline to import images and labels from the Caltech-101
dataset. The code resides in `data/caltech_101.py` and features an
interface similar to the already existing MNIST equivalent. Images
are converted to greyscale, cropped to a center-square and rescaled
to a fixed resolution.

2. a Jupyter Notebook in which a CapsNet is trained to distinguish
images of human faces from images of motorcycles. Motorcycles have
been chosen as they are quite prominently featured in the Caltech-101
dataset and also reduce the possibility of accidentally contianing
a face-like shape.

The trained model achieves an accuracy of 98% on the test set.